♻️ REFACTOR: 어드민 타입 분리 #165
